Rename package to building_metrics_3d
