+++ pyproject.toml
@@ -20,6 +20,7 @@ dependencies = [
     "miniball==1.2.0",
     "numpy==1.24.4",
     "pandas==2.0.3",
+    "pymesh",
     "pymeshfix==0.16.2",
     "python-polylabel",
     "pyvista==0.36.1",
