+++ src/building_metrics_3d/cityCompare.py
@@ -2,8 +2,8 @@ import click
 import json
 import numpy as np
 from tqdm import tqdm
-from urban_morphology_3d import cityjson
-from urban_morphology_3d.helpers.mesh import difference, symmetric_difference, to_pymesh, to_pyvista, intersect
+from building_metrics_3d import cityjson
+from building_metrics_3d.helpers.mesh import difference, symmetric_difference, to_pymesh, to_pyvista, intersect
 import pyvista as pv
 import pandas as pd
 import pymesh
