+++ src/building_metrics_3d/cityjson.py
@@ -2,7 +2,7 @@
 
 import numpy as np
 from shapely.geometry import Polygon, MultiPolygon
-from urban_morphology_3d.helpers.geometry import project_2d, surface_normal, triangulate, triangulate_polygon
+from building_metrics_3d.helpers.geometry import project_2d, surface_normal, triangulate, triangulate_polygon
 import pyvista as pv
 
 def get_surface_boundaries(geom):
