+++ src/building_metrics_3d/helpers/geometry.py
