+++ src/building_metrics_3d/helpers/mesh.py
