+++ src/building_metrics_3d/helpers/minimumBoundingBox.py
