+++ src/building_metrics_3d/helpers/smallestenclosingcircle.py
